backend/requirements.txt

# Backend requirements
# Install with: pip install -r requirements.txt

# Core dependencies
fastapi>=0.104.1
uvicorn>=0.24.0
flask>=3.0.0
flask-cors>=4.0.0
python-multipart>=0.0.6
python-dotenv>=1.0.0

# Image processing
Pillow>=10.1.0
transformers>=4.35.2
torch>=2.2.1
torchvision>=0.17.1

# LLM integration
openai>=1.3.0
groq

# Emoji processing
emoji>=2.1.0

# Utilities
requests>=2.31.0
pydantic>=2.5.0

# Development (optional)
pytest>=7.4.3
pytest-asyncio>=0.21.1
black>=23.12.0
flake8>=6.1.0
